More checks before the rpm Koji build task
The zuul_rpm_publish will fail if the NVR already exist
on Koji.

This patch first check the NVR we want to build is not
already part of the target tag. If that the case it exits 0.
This avoid useless failures in the gate.

Else the build is done as usual but in case of failure as a fallback
we try a tag-build to the targeted tag. This is useful
when a patch is backported from master to a release branch but
the generated NVR is the same and have been already published on Koji.
In that case the publish_rpm needs to try to tag-build.

This is an issue that can be seen here:
https://softwarefactory-project.io/jenkins/job/sf-rpm-publish/338/console

Change-Id: I2b217256da3e8dc5406c2ba2d45ad02b4d37ae5c