Update patch set 2
Patch Set 2:

I think we need to have this one validated and merged first because that is the first part of our CI where the packages are built.
I know that sfinfo is a config repo making it hard to test speculatively. But we should be able to attach the sf-rpm-build job via an untrusted repo and in the experimental pipeline. I think will see that most of the deps you added in sf-ci/install_sf_ci_packages.yml are need for rpmbuild and rpmpublish. So we will need to improve rpmbuild.yaml and rpmpublish.yaml to install the right deps to run our packaging tooling.

Patch-set: 2
Label: Code-Review=0